Added session2.py from files
This code prints out the sum, difference, and product of two numbers defined by x and y. Some x and y values carry a floating point number while others carry a single integer. The output prints what x and y are defined as, the sum/difference/product of the two, and also lists the data type that the end result is.