Added session2.py program file
This program runs a defined x and y value of an integer or floating point number. The output prints out what the predefined x and y variables were defined as, while printing the sum/difference/product of the x and y values. The output also prints out the data type of each answer.