Added Session-8 program files
Create a Jupyter notebook, import matplotlib. Write cells that create an array x ranging from [0,1] in 100 steps and that defines functions that return sin(x) and cos(x).  In a new cell use to create a multipanel plot (1 row, 2 columns), plotting sin(x) vs. x in the left panel and cos(x) vs. x in the right panel. Label the panels with sin(x) and cos(x), and save the figure as a PDF.